replace: app router(top page) -> top
